Renamed package game to board. Was doing some sealed class testing to check if it would be good choice for using as Game data/rules instead of individual ViewModels. Kept the test file in as reference.
