+++ app/src/main/java/com/heyzeusv/solitaire/ui/board/BoardLayout.kt
@@ -1,4 +1,4 @@
-package com.heyzeusv.solitaire.ui.game
+package com.heyzeusv.solitaire.ui.board
 
 import androidx.compose.animation.core.animate
 import androidx.compose.animation.core.tween
