+++ app/src/main/java/com/heyzeusv/solitaire/ui/board/Tableau.kt
@@ -1,4 +1,4 @@
-package com.heyzeusv.solitaire.ui.game
+package com.heyzeusv.solitaire.ui.board
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
