+++ app/src/test/java/com/heyzeusv/solitaire/ui/AlaskaViewModelTest.kt
@@ -1,8 +1,8 @@
 package com.heyzeusv.solitaire.ui
 
 import com.heyzeusv.solitaire.data.ShuffleSeed
-import com.heyzeusv.solitaire.ui.game.AlaskaViewModel
-import com.heyzeusv.solitaire.ui.game.YukonViewModel
+import com.heyzeusv.solitaire.ui.board.AlaskaViewModel
+import com.heyzeusv.solitaire.ui.board.YukonViewModel
 import com.heyzeusv.solitaire.ui.scoreboard.ScoreboardViewModel
 import com.heyzeusv.solitaire.util.TestCards
 import kotlinx.coroutines.Dispatchers
