+++ app/src/test/java/com/heyzeusv/solitaire/ui/ClassicWestcliffViewModelTest.kt
@@ -3,7 +3,7 @@ package com.heyzeusv.solitaire.ui
 import com.heyzeusv.solitaire.data.Card
 import com.heyzeusv.solitaire.data.PileHistory
 import com.heyzeusv.solitaire.data.ShuffleSeed
-import com.heyzeusv.solitaire.ui.game.ClassicWestcliffViewModel
+import com.heyzeusv.solitaire.ui.board.ClassicWestcliffViewModel
 import com.heyzeusv.solitaire.ui.scoreboard.ScoreboardViewModel
 import com.heyzeusv.solitaire.util.ResetOptions
 import com.heyzeusv.solitaire.util.TestCards
