+++ app/src/test/java/com/heyzeusv/solitaire/ui/EasthavenViewModelTest.kt
@@ -3,7 +3,7 @@ package com.heyzeusv.solitaire.ui
 import com.heyzeusv.solitaire.data.Card
 import com.heyzeusv.solitaire.data.PileHistory
 import com.heyzeusv.solitaire.data.ShuffleSeed
-import com.heyzeusv.solitaire.ui.game.EasthavenViewModel
+import com.heyzeusv.solitaire.ui.board.EasthavenViewModel
 import com.heyzeusv.solitaire.ui.scoreboard.ScoreboardViewModel
 import com.heyzeusv.solitaire.util.ResetOptions
 import com.heyzeusv.solitaire.util.TestCards
